requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
beautifulsoup4==4.13.4
fastembed==0.5.1
feedparser==6.0.11
apache-airflow-providers-weaviate==3.0.3
llama-index==0.12.42
llama-index-embeddings-fastembed==0.3.2
llama-index-embeddings-huggingface==0.5.4
llama-index-vector-stores-weaviate==1.3.1
